Added options for local smtp server and git repo
Added an option to install a  local smtp server (exim4 in this case):
redmine (email.yml) will need to be setup to use the local server.

Added an option to clone a github repo to allow redmine projects to have a
local repos to browse. For now only one repo is setup. A cronjob has been
created to sync the repo every few minutes.

Added the SMTP instructions in the README