Information for Eucalyptus provided images
The images are still in testing mode, but they will be published very
soon. They come in 2 kind: with a small root files system (to be used with
unmodified Eucalyptus 2 deployment) and a larger (4GB) root file systems
(the Cloud Administrator will need to change the defaults for the image
types).